requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
cryptography==44.0.1
Django==5.1.6
django-crispy-forms==2.3
django-microsoft-auth==3.0.1
idna==3.10
msal==1.31.1
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
